make max reorg height a signed int

+++ src/main.h
@@ -265,7 +265,7 @@ CAmount GetFluxnodeSubsidy(int nHeight, const CAmount& blockValue, int nNodeTier
 CAmount GetExchangeFundAmount(int nHeight, const Consensus::Params& consensusParams);
 CAmount GetFoundationFundAmount(int nHeight, const Consensus::Params& consensusParams);
 bool IsSwapPoolInterval(const int64_t nHeight);
-unsigned int GetMaxReorgDepth(const int64_t nHeight);
+int64_t GetMaxReorgDepth(const int64_t nHeight);
 
 /**
  * Prune block and undo files (blk???.dat and undo???.dat) so that the disk space used is less than a user-defined target.
